+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v0.3.0 (Jan 15, 2025)
 
 * More obvious styling for the "add new" template option on Home Screen
+* Allow switching of the template directly on the timer screen